Clean up the installation instructions.
Don't assume what the path will be on Unix-like systems; have the user
find the path from Help -> About Wireshark first.  It is *not*
guaranteed to be in ~/.wireshark; Wireshark looks there for backwards
compatibility, but we moved the default location to ~/.config/wireshark
- see Wireshark bug 6353.

Don't assume the directory will exist; give the user instructions to
create it, on Unix-like systems, if it doesn't.

Speak of Unix-like systems in general, not just Linux or OS X - mention
*BSD, Solaris, AIX, and HP-UX as other examples.  (Feel free to add
no-longer-actively-supported UN*Xes such as IRIX or Tru64 UNIX if you
wish.)

Don't mention cloning up front - treat it and "download the zip file" as
equal possibilities, although we still mention cloning before
downloading the zip file.